pyproject.toml
[tool.poetry]
name = "secret-harvest"
version = "0.2.2"
description = ""
authors = ["Robert Thomas"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
git-python = "^1.0.3"
rich = "^14.0.0"
tabulate = "^0.9.0"

[tool.poetry.scripts]
secret-harvest = "secret_harvest.main:main_with_args"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
